Create README.md
Playing a game with paper money can be difficult. It can be hard to determine how much money you have by looking at the stack in front of you. This leaves the player counting their money several times during the game, which slows down play for everyone.

To alleviate this, many avid gamers bring along poker chips. Due to the compact size and color coding, counting money is far more efficient. However, this solution is also a burden. Poker chip sets are bulky and heavy to carry around.

Also, this method does nothing to mitigate the time lost making change with the bank. Players who are used to engaging in games with money are familiar with this dynamic. To understand how much money the player has, it is most efficient to have as few pieces of paper or chips as possible. However, when making transactions, it is far more efficient to have enough change to make transactions without going to the bank. A balance is needed, and is often hard to find.

Another issue with money systems in games is making sure that these transactions are occurring. In a game like monopoly, keeping track of these transactions is simple, since they are completed on a turn-by-turn basis. However, in an 18XX game like 1830, keeping track of these transactions is far more difficult. Players are often left trying to figure out if they remembered to pay themselves or not.

To solve all these problems, the proposed program keeps track of the player’s money for them. The proposed program is a phone app. The majority of the display will be committed to a large print currency amount. Below it will be a list of all the transactions that have occurred during play, much like a bank statement. This will help alleviate the issue of missing transactions. If a player is wondering if a transaction occurred, they would simply look into their history to know for sure. This history is created when the player makes a transaction. The user enters an amount to adjust their total by, and then commits the change.